Refactoring web module
