+++ spring-survey-web/src/main/java/hu/unideb/inf/survey/web/question/model/QuestionDto.java
@@ -1,4 +1,6 @@
-package hu.unideb.inf.survey.web.survey.model;
+package hu.unideb.inf.survey.web.question.model;
+
+import hu.unideb.inf.survey.web.answer.model.AnswerDto;
 
 import java.util.List;
