+++ spring-survey-web/src/main/java/hu/unideb/inf/survey/web/survey/controller/CheckSurveyController.java
@@ -3,9 +3,9 @@ package hu.unideb.inf.survey.web.survey.controller;
 import hu.unideb.inf.survey.service.SelectedAnswerService;
 import hu.unideb.inf.survey.service.SurveyQuestionService;
 import hu.unideb.inf.survey.service.domain.SurveyQuestionDomain;
-import hu.unideb.inf.survey.web.survey.model.QuestionStatisticDto;
+import hu.unideb.inf.survey.web.question.model.QuestionStatisticDto;
 import hu.unideb.inf.survey.web.survey.model.SurveyDataDto;
-import hu.unideb.inf.survey.web.survey.transformer.QuestionStatisticDtoTransformer;
+import hu.unideb.inf.survey.web.question.transformer.QuestionStatisticDtoTransformer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
