+++ spring-survey-web/src/main/java/hu/unideb/inf/survey/web/survey/controller/EditSurveyController.java
@@ -4,10 +4,10 @@ import hu.unideb.inf.survey.service.SurveyQuestionService;
 import hu.unideb.inf.survey.service.SurveyService;
 import hu.unideb.inf.survey.service.domain.SurveyDomain;
 import hu.unideb.inf.survey.service.domain.SurveyQuestionDomain;
-import hu.unideb.inf.survey.web.survey.model.QuestionDto;
+import hu.unideb.inf.survey.web.question.model.QuestionDto;
 import hu.unideb.inf.survey.web.survey.model.SurveyDto;
 import hu.unideb.inf.survey.web.survey.transformer.SurveyDtoTransformer;
-import hu.unideb.inf.survey.web.survey.transformer.SurveyQuestionDtoTransformer;
+import hu.unideb.inf.survey.web.question.transformer.SurveyQuestionDtoTransformer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
