+++ spring-survey-web/src/main/java/hu/unideb/inf/survey/web/survey/model/SurveyDataDto.java
@@ -1,5 +1,7 @@
 package hu.unideb.inf.survey.web.survey.model;
 
+import hu.unideb.inf.survey.web.question.model.QuestionStatisticDto;
+
 import java.util.List;
 
 public class SurveyDataDto {
